+++ src/CommerceKitServiceProvider.php
@@ -7,6 +7,7 @@ namespace Lalalili\CommerceKit;
 use Lalalili\CommerceKit\Contracts\CartDiscountRefresher;
 use Lalalili\CommerceKit\Coupons\CouponCartConditionFactory;
 use Lalalili\CommerceKit\Pipelines\CartDiscountRefreshPipeline;
+use Lalalili\CommerceKit\Recurring\RecurringCheckoutContextBuilder;
 use Lalalili\CommerceKit\Support\ConfiguredCartDiscountRefresher;
 use Lalalili\CommerceKit\Support\ConfiguredCheckoutOrderBuilder;
 use Lalalili\CommerceCore\Contracts\CheckoutOrderBuilder;
@@ -28,5 +29,6 @@ class CommerceKitServiceProvider extends PackageServiceProvider
         $this->app->singletonIf(CheckoutOrderBuilder::class, ConfiguredCheckoutOrderBuilder::class);
         $this->app->singleton(CouponCartConditionFactory::class);
         $this->app->singleton(CartDiscountRefreshPipeline::class);
+        $this->app->singleton(RecurringCheckoutContextBuilder::class);
     }
 }